Clarify the difference between a tunnel and a relay
It took me a while to realize that a relay it's not just
tunneling a stream to a destination through a proxy peer
but rather it is making a destination peer
aware that the source peer is reachable through the relaying
node. Hopefully this makes that more explicit.

License: MIT
Signed-off-by: Hector Sanjuan <code@hector.link>