noise-libp2p spec: rethink message data construction.
The current message data construction is not secure enough, as it is
vulnerable to replay attacks by a MITM agent that records previous
handshakes, and transplants old user data from those handshakes into new
ones, as long as the static key hasn't changed.

We fix this vulnerability by introducing a signature over the whole
message payload against the ephemeral session key. This "seals" the
payload so that it's only valid for that exchange.

This PR simplifies protobuf field naming.

Finally, we formalise in which Noise messages of IK and XX the message
payload is to be shared, to guarantee secrecy, integrity and
authentication.