noise-libp2p: introduce "handshake seal"; more.
The current construction is vulnerable to replay attacks by a MITM agent
that records previous handshakes, and transplants old user data from
those handshakes into new ones, as long as the static key hasn't changed.

We fix this vulnerability by leveraging Noise channel binding, in order
to "seal" the handshake and assert that the exchanged data was witnessed
equally by both parties.

This commit also simplifies protobuf field naming.

Finally, we formalise in which Noise messages of IK and XX the message
payload is to be shared, to guarantee secrecy, integrity and
authentication.