Remove second-guessing naming of Message field
Small tweak, but it kept jumping out at me as I'm reading the spec. A spec should be written in an authoritative manner and not second-guess itself. Right now it reads more like a note to self that was left over.

If this is a possible enhancement/renaming, it should be captured in a separate issue IMO.

It is also unlikely that this field will change, considering how many clients implemented it with this naming.

Finally, I personally think the name is perfectly fine :)