rendezvous/: Clarify `UNREGISTER` behaviour and remove `id` (#348)
* Remove `id` field from `Unregister` message

All connections in libp2p are authenticated. As such, we don't have
to include the PeerId in the `Unregister` message. We only allow
peers to unregister themselves and therefore, this `id` would always
be equal to the one we are learn from the authentication layer. Having
to compare those and ensure they are equal is an unnecessary error
path.

To be backwards-compatible with existing deployments of the rendezvous
protocol, the field is only commented out. If this record ever gets
extended, the next field should use id `3`.

* Clarify absence of `UNREGISTER` response