Announcesub spec
Announcesub is a new pubsub protocol over libp2p, heavily based on gossipsub.

The protocol is designed in a way to reduce the number of duplicates to zero,
trading off with more latency.

It is actually the same as Gossipsub except that instead of sending the
messages right away to mesh peers, it sends IANNOUNCE instead and waits
for INEED before sending the actual messages.

This guarantees that each node will receives exactly one copy of
messages.