Gossipsub v2.0 spec
gossipsub v2.0 allows you to reduce the number of duplicates lower or to
zero trading off with more latency.

It works by probabilistically deciding to forward the message eargerly
or lazily to mesh peers.

If it decides to send eargerly, it just forwards the message right away.

If it decides to send lazily, it sends IANNOUNCE instead and waits for
INEED before sending the actual messages.

Notice that if the probability is configured to be 1, it guarantees that
each node will receives exactly one copy of messages, which means no
duplicates.